Update main.c to enable miniUART on RPI3 B+
The baud rate of miniUART relies on the core clock frequency. RPI3 B+ changed the frequency from 250MHz to 400MHz and varies between them. This makes the baud rate calculation difficult. So the rate is fixed to one rate (RPI3 400MHz, RPI2 250MHz) using the mailbox channel to the video core.